refactor: intcode uses i128

+++ src/amplification_circuit.rs
@@ -2,26 +2,26 @@ use std::collections::HashSet;
 
 use super::intcode::Computer;
 
-fn all_unique_settings(x: &Vec<isize>) -> bool {
-    let items: HashSet<isize> = x.into_iter().cloned().collect();
+fn all_unique_settings(x: &Vec<i128>) -> bool {
+    let items: HashSet<i128> = x.into_iter().cloned().collect();
     items.len() == x.len()
 }
 
-fn run_amplifiers(program: &Vec<isize>, settings: &Vec<isize>) -> isize {
-    let mut output: Vec<isize> = vec![0];
+fn run_amplifiers(program: &Vec<i128>, settings: &Vec<i128>) -> i128 {
+    let mut output: Vec<i128> = vec![0];
     for &setting in settings {
-        let input = vec![setting as isize, *output.first().unwrap()];
+        let input = vec![setting, *output.first().unwrap()];
         let computer = Computer::new(program.clone(), &input);
         output = computer.collect();
     }
     *output.first().unwrap()
 }
 
-fn run_amplifiers_with_feedback(program: &Vec<isize>, settings: &Vec<isize>) -> isize {
+fn run_amplifiers_with_feedback(program: &Vec<i128>, settings: &Vec<i128>) -> i128 {
     let mut computers: Vec<Computer> = settings.iter()
         .map(|&setting| Computer::new(program.clone(), &[setting]))
         .collect();
-    let mut last_output: isize = 0;
+    let mut last_output: i128 = 0;
     'outer: loop {
         for computer in computers.iter_mut() {
             computer.input_queue.push_back(last_output);
@@ -35,7 +35,7 @@ fn run_amplifiers_with_feedback(program: &Vec<isize>, settings: &Vec<isize>) -> 
     last_output
 }
 
-fn maximum_input_combination(program: &Vec<isize>) -> isize {
+fn maximum_input_combination(program: &Vec<i128>) -> i128 {
     iproduct!(0..5, 0..5, 0..5, 0..5, 0..5)
         .into_iter()
         .map(|(a, b, c, d, e)| vec![a, b, c, d, e])
@@ -45,7 +45,7 @@ fn maximum_input_combination(program: &Vec<isize>) -> isize {
         .unwrap()
 }
 
-fn maximum_input_combination_feedback(program: &Vec<isize>) -> isize {
+fn maximum_input_combination_feedback(program: &Vec<i128>) -> i128 {
     iproduct!(5..10, 5..10, 5..10, 5..10, 5..10)
         .into_iter()
         .map(|(a, b, c, d, e)| vec![a, b, c, d, e])
@@ -55,12 +55,12 @@ fn maximum_input_combination_feedback(program: &Vec<isize>) -> isize {
         .unwrap()
 }
 
-pub fn solve_part_one() -> isize {
+pub fn solve_part_one() -> i128 {
     let program = Computer::get_program(7);
     maximum_input_combination(&program)
 }
 
-pub fn solve_part_two() -> isize {
+pub fn solve_part_two() -> i128 {
     let program = Computer::get_program(7);
     maximum_input_combination_feedback(&program)
 }
@@ -71,7 +71,7 @@ mod tests {
 
     #[test]
     fn test_maximum_input_combination() {
-        let values: Vec<(Vec<isize>, isize)> = vec![
+        let values: Vec<(Vec<i128>, i128)> = vec![
             (vec![3, 15, 3, 16, 1002, 16, 10, 16, 1, 16, 15, 15, 4, 15, 99, 0, 0], 43210),
             (vec![3, 23, 3, 24, 1002, 24, 10, 24, 1002, 23, -1, 23, 101, 5, 23, 23, 1, 24, 23, 23, 4, 23, 99, 0, 0], 54321),
             (vec![3, 31, 3, 32, 1002, 32, 10, 32, 1001, 31, -2, 31, 1007, 31, 0, 33, 1002, 33, 7, 33, 1, 33, 31, 31, 1, 32, 31, 31, 4, 31, 99, 0, 0, 0], 65210),
@@ -90,7 +90,7 @@ mod tests {
 
     #[test]
     fn test_maximum_input_combination_feedback() {
-        let values: Vec<(Vec<isize>, isize)> = vec![
+        let values: Vec<(Vec<i128>, i128)> = vec![
             (vec![3, 26, 1001, 26, -4, 26, 3, 27, 1002, 27, 2, 27, 1, 27, 26, 27, 4, 27, 1001, 28, -1, 28, 1005, 28, 6, 99, 0, 0, 5], 139629729),
             (vec![3, 52, 1001, 52, -5, 52, 3, 53, 1, 52, 56, 54, 1007, 54, 5, 55, 1005, 55, 26, 1001, 54, -5, 54, 1105, 1, 12, 1, 53, 54, 53, 1008, 54, 0, 55, 1001, 55, 1, 55, 2, 53, 55, 53, 4, 53, 1001, 56, -1, 56, 1005, 56, 6, 99, 0, 0, 0, 0, 10], 18216),
         ];
